commit type: chore
In this commit we are removing dependency pinning because, to be honest there are lot of bugs in web code and so its best to stay at the newest release.
Additionally, some of the dependencies break when trying to pip install them, so I just removed them so the build pipeline doesn't break.

No issues tracked