commit type: refactor
The current style of the application folder doesn't lend itself to a scalable structure.  This will be replaced in future commits with the easier to type 'app' folder, as well as a clearler delination between front end code and backend code, to be more inline with the MVC style framework.

no issues tracked