commit type: refactor
Removing unused code.  The style of the html can be more flexible now that we've moved to an MVC framework.

At the present moment the templates are taking on too much work, which will likely be moved over to JS (probably react).  So the unnecessary elements have been removed

no issues tracked